Grabs the 'Default' network from the list of networks
If more than one network is provided it will look for the one that is
defined as being the default gateway. If none are defined to contain the
default gateway then a random network is chosen.  If there is only one
it is returned if there are none then an empty network is returned
[#94624368]
Signed-off-by: Hua Zhang <zhuadl@cn.ibm.com>