Disable chrony in warden containers
During agent bootstrap it runs /var/vcap/bosh/bin/sync-time
Which executes `chronyc waitsync 10` which in turn hangs for 90s
because it can not reach the chronyd daemon. This daemon is not
running because there is no systemd in containers.

Long story short, we actually don't need to sync time in a container
since it shares the time with the host (can only change the time when
running priveleged).

The solution is to comment out invokations of chronyc in the sync-time
script during container start.