🐞 Enable the mounting of the 9th loopback device
We are seeing "failed to setup loop device" when attaching persistent disks
to "VMs"

strace was showing that without the loop-control device, the losetup would iterate
through the entries in /sys/dev/block/7:* to determine if a loopback device was in use;
However, there were no entries in that directory for devices 7:8 and beyond.

This caused mount to fail when it called down into losetup to configure the loopback
device for the disk.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>